requirements.txt
numpy
matplotlib
scipy
scikit-learn
torch
torchvision
PyYAML
tqdm
